Implemented POC PDM-TO-PCM conversion logic. Not works as expected.

+++ guitar_tuner/src/signal_buffer.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-void push_to_signal_buffer(int32_t sample);
+void push_to_signal_buffer(int16_t sample);
 
 #endif
